Define dstime as int64_t and use std::chrono for m_clock_getmonotonictime
Remove warinings

+++ src/win32/waiter.cpp
@@ -54,7 +54,12 @@ int WinWaiter::wait()
     if (index <= MAXIMUM_WAIT_OBJECTS)
     {
         assert(!handles.empty());
-        DWORD dwWaitResult = WaitForMultipleObjectsEx((DWORD)index, &handles.front(), FALSE, maxds * 100, TRUE);
+        DWORD dwWaitResult = WaitForMultipleObjectsEx(static_cast<DWORD>(index),
+                                                      &handles.front(),
+                                                      FALSE,
+                                                      static_cast<DWORD>(maxds * 100),
+                                                      TRUE);
+
         assert(dwWaitResult != WAIT_FAILED);
 
 #ifdef MEGA_MEASURE_CODE
